Fix step function example
